Improved precision of code

+++ hdr/header.hpp
@@ -4,5 +4,5 @@ double calculate_satellite_distance(double angle, double radius_earth, double ra
 double calculate_Cn2(double x, double angle, double radius_earth, double satellite_distance, double wind_speed, double Cn2_0);
 double integrate_turbulence(double angle, double radius_earth, double satellite_distance, double wind_speed, double Cn2_0);
 double calculate_height(double angle, double radius_earth, double distance);
-float calculate_beam_intensity(double satellite_distance, double integrated_turbulence, double diameter_laser, double wavenumber_laser, double curvature_laser, float inner_scale_size);
+double calculate_beam_intensity(double satellite_distance, double integrated_turbulence, double diameter_laser, double wavenumber_laser, double curvature_laser, float inner_scale_size);
 #endif
